card_bot/requirements.txt
python-telegram-bot==20.7
pytesseract==0.3.10
Pillow==10.1.0
cryptography==41.0.7
